+++ index.js
@@ -1,5 +1,7 @@
+'use strict';
+
 /**
   The entry point.
   @module Sequelize
 **/
-module.exports = require("./lib/sequelize");
+module.exports = require('./lib/sequelize');