+++ test/integration/model/upsert.test.js
@@ -12,6 +12,10 @@ var chai = require('chai')
   , current = Support.sequelize;
 
 describe(Support.getTestDialectTeaser('Model'), function() {
+  before(function () {
+    this.clock = sinon.useFakeTimers();
+  });
+
   beforeEach(function() {
     this.User = this.sequelize.define('user', {
       username: DataTypes.STRING,
@@ -31,7 +35,24 @@ describe(Support.getTestDialectTeaser('Model'), function() {
       blob: DataTypes.BLOB
     });
 
+    this.ModelWithFieldPK = this.sequelize.define('ModelWithFieldPK', {
+      userId: {
+        field: 'user_id',
+        type: Sequelize.INTEGER,
+        autoIncrement: true,
+        primaryKey: true
+      },
+      foo:{
+        type: DataTypes.STRING,
+        unique: true
+      }
+    });
+
     return this.sequelize.sync({ force: true });
+  });
+
+  after(function () {
+    this.clock.restore();
   });
 
   if (current.dialect.supports.upserts) {
@@ -44,9 +65,8 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
-            return this.User.upsert({ id: 42, username: 'doe' });
-          });
+          this.clock.tick(1000);
+          return this.User.upsert({ id: 42, username: 'doe' });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;
@@ -70,9 +90,8 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
-            return this.User.upsert({ foo: 'baz', bar: 19, username: 'doe' });
-          });
+          this.clock.tick(1000);
+          return this.User.upsert({ foo: 'baz', bar: 19, username: 'doe' });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;
@@ -136,10 +155,10 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created2).to.be.ok;
           }
 
-          return Promise.delay(1000).bind(this).then(function() {
-            // Update the first one
-            return User.upsert({ a: 'a', b: 'b', username: 'doe' });
-          });
+
+          this.clock.tick(1000);
+          // Update the first one
+          return User.upsert({ a: 'a', b: 'b', username: 'doe' });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;
@@ -183,9 +202,9 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
-            return this.User.upsert({ id: 42, username: 'doe', blob: new Buffer('andrea') });
-          });
+
+          this.clock.tick(1000);
+          return this.User.upsert({ id: 42, username: 'doe', blob: new Buffer('andrea') });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;
@@ -210,9 +229,7 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
-            return this.User.upsert({ id: 42, baz: 'oof' });
-          });
+          return this.User.upsert({ id: 42, baz: 'oof' });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;
@@ -226,6 +243,30 @@ describe(Support.getTestDialectTeaser('Model'), function() {
         });
       });
 
+      it('works with primary key using .field', function () {
+        return this.ModelWithFieldPK.upsert({ userId: 42, foo: 'first' }).bind(this).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).to.be.undefined;
+          } else {
+            expect(created).to.be.ok;
+          }
+
+
+          this.clock.tick(1000);
+          return this.ModelWithFieldPK.upsert({ userId: 42, foo: 'second' });
+        }).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).to.be.undefined;
+          } else {
+            expect(created).not.to.be.ok;
+          }
+
+          return this.ModelWithFieldPK.findOne({ userId: 42 });
+        }).then(function(instance) {
+          expect(instance.foo).to.equal('second');
+        });
+      });
+
       it('works with database functions', function() {
         return this.User.upsert({ id: 42, username: 'john', foo: this.sequelize.fn('upper', 'mixedCase1')}).bind(this).then(function(created) {
           if (dialect === 'sqlite') {
@@ -233,9 +274,10 @@ describe(Support.getTestDialectTeaser('Model'), function() {
           } else {
             expect(created).to.be.ok;
           }
-          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
-            return this.User.upsert({ id: 42, username: 'doe', foo: this.sequelize.fn('upper', 'mixedCase2') });
-          });
+
+
+          this.clock.tick(1000);
+          return this.User.upsert({ id: 42, username: 'doe', foo: this.sequelize.fn('upper', 'mixedCase2') });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).to.be.undefined;